Refactor linked bundle workspace modules
